multiworkspace implemented

+++ app/Http/Controllers/Features/Product/InvitationController.php
@@ -2,18 +2,18 @@
 
 namespace App\Http\Controllers\Features\Product;
 
-use App\Models\User;
-use App\Models\Product;
-use App\Models\Invitation;
-use Illuminate\Support\Str;
-use Illuminate\Http\Request;
-use App\Models\Scopes\OwnerScope;
-use Spatie\Permission\Models\Role;
-use App\Services\InvitationService;
 use App\Http\Controllers\Controller;
+use App\Http\Requests\TeamInvitationRequest;
+use App\Models\Invitation;
+use App\Models\Product;
+use App\Models\Scopes\OwnerScope;
+use App\Models\User;
+use App\Services\InvitationService;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Hash;
-use App\Http\Requests\TeamInvitationRequest;
+use Illuminate\Support\Str;
+use Spatie\Permission\Models\Role;
 
 class InvitationController extends Controller
 {
@@ -70,6 +70,9 @@ class InvitationController extends Controller
      */
     public function accept($token)
     {
+        Auth::guard('web')->logout();
+        request()->session()->invalidate();
+        request()->session()->regenerateToken();
         $invitation = Invitation::withoutGlobalScope(OwnerScope::class)->where('token', $token)->first();
         $user       = User::where('email', $invitation->email)->first();
         $id         = base64_encode($invitation->id);
@@ -77,7 +80,7 @@ class InvitationController extends Controller
         if ($user) {
             if ($user->owner_id != $invitation->owner_id) {
                 $newUser = User::create([
-                    'name'              => $invitation->first_name ." ". $invitation->last_name,
+                    'name'              => $invitation->first_name . " " . $invitation->last_name,
                     'email_verified_at' => now(),
                     'password'          => Hash::make(Str::random()),
                     'first_name'        => $invitation->first_name,
@@ -93,9 +96,6 @@ class InvitationController extends Controller
                 notify()->success(trans('A new workspace has been created!'));
                 return to_route('dashboard');
             }
-            Auth::guard('web')->logout();
-            request()->session()->invalidate();
-            request()->session()->regenerateToken();
             Auth::login($user);
             $this->assignToUser($invitation, $user);
             notify()->success('You have already accepted the invitation!');
@@ -130,7 +130,7 @@ class InvitationController extends Controller
             return redirect()->route('home');
         }
         $user = User::create([
-            'name'              => $invitation->first_name . " ". $invitation->last_name,
+            'name'              => $invitation->first_name . " " . $invitation->last_name,
             'email'             => $invitation->email,
             'email_verified_at' => now(),
             'password'          => Hash::make($request->password),
@@ -174,17 +174,21 @@ class InvitationController extends Controller
         return redirect()->route('invitation.index')->with('success', 'Invitation deleted successfully');
     }
 
-    public function assignToUser(Invitation $invitation, User $user): void
+    protected function assignToUser(Invitation $invitation, User $user): void
     {
-        $user->myProducts()->detach();
-        foreach ($invitation->products as $product) {
-            $user->myProducts()->attach($product->id);
-            $invitation->products()->updateExistingPivot($product->id, [ 'is_accepted' => true ]);
+        if ($invitation->products) {
+            $user->myProducts()->detach();
+            foreach ($invitation->products as $product) {
+                $user->myProducts()->attach($product->id);
+                $invitation->products()->updateExistingPivot($product->id, [ 'is_accepted' => true ]);
+            }
         }
 
-        $user->tasks()->detach();
-        foreach ($invitation->tasks as $task) {
-            $user->tasks()->attach($task);
+        if ($invitation->tasks) {
+            $user->tasks()->detach();
+            foreach ($invitation->tasks as $task) {
+                $user->tasks()->attach($task);
+            }
         }
         $user->owner_id = $invitation->owner_id;
         $user->type     = 'member';
